requirements.txt
# Pin major versions for reproducible deploys; bump after testing.
Django>=6.0,<7
djangorestframework>=3.15,<4
djangorestframework-simplejwt>=5.3,<6
django-cors-headers>=4.0,<5
gunicorn>=22,<24
psycopg[binary]>=3.2,<4
dj-database-url>=2.1,<3
whitenoise>=6.6,<7
